Basic Async MySQL Client Support
Summary:
WebScaleSQL Feature: Async Client

This diff builds loosely upon the 5.1 client work at Facebook, and
brings async to 5.6.  It is not fully feature complete (no SSL. no
compression, only supports the default native auth) but it is a good
first step, is sufficient for our needs, and we can extend it.

This also adds a #define to indicate the nonblocking client is available.

This changes slightly the ABI and so it also renames 'libmysqlclient' to
'libwebscalesqlclient' -- however, this is compatible with almost every
use case (changes are additive only, but do include changes to some data
structures).

Test Plan: mtr, testing against internal tools

Reviewers: chip, inaam-rana, pivanof, darnaut, jeremycole

Reviewed By: jeremycole, inaam-rana, pivanof, chip

Subscribers: fe, darnaut, methane, steaphan, atcurtis, chip, liang.guo.752, andrew-ford, jeremycole, flamingcow, MarkCallaghan, jtolmer

Differential Revision: https://reviews.facebook.net/D17031
Differential Revision: https://reviews.facebook.net/D33777
Differential Revision: https://reviews.facebook.net/D33801